remove remaining nil comparison
